Refactored files and split into folders
